+++ app.go
@@ -80,7 +80,7 @@ func (a *App) run() {
 	rl.BeginDrawing()
 	rl.ClearBackground(rl.Black)
 	rl.BeginMode2D(a.cam)
-	a.drawViewport()
+	a.drawSceneViewport()
 	a.control.Draw(a)
 	a.drawDialogs()
 	rl.EndMode2D()
