+++ script.go
@@ -113,40 +113,3 @@ func (m Method) ForEach(f func(string)) {
 func (m Method) String() string {
 	return strings.Join(m, ".")
 }
-
-// ScriptRun is a command to run a script.
-type ScriptRun struct {
-	ScriptRef ResourceRef
-}
-
-func (c ScriptRun) Execute(app *App, prom *Promise) {
-	script, ok := app.scripts[c.ScriptRef]
-	if ok {
-		prom.CompleteWithValue(script)
-		return
-	}
-
-	// The script was not loaded yet, so we load and execute it now.
-	script = app.res.LoadScript(c.ScriptRef)
-	if script == nil {
-		log.Panicf("Script not found: %s", c.ScriptRef)
-	}
-	script.init(app, c.ScriptRef)
-	app.scripts[c.ScriptRef] = script
-	script.run(app, prom)
-}
-
-// ScriptCall is a command to call a script function.
-type ScriptCall struct {
-	ScriptRef ResourceRef
-	Method    Method
-}
-
-func (c ScriptCall) Execute(app *App, prom *Promise) {
-	script, ok := app.scripts[c.ScriptRef]
-	if !ok {
-		log.Panicf("Script not found: %s", c.ScriptRef)
-	}
-
-	prom.Bind(script.Call(c.Method))
-}
